// src/@v2/forms/domain/enums/form-copsoq-category.enum.ts
export enum FormCOPSOQCategoryEnum {
  COMPORTAMENTOS_OFENSIVOS = 'COMPORTAMENTOS_OFENSIVOS',
  EXIGENCIAS_LABORAIS = 'EXIGENCIAS_LABORAIS',
  INTERFACE_TRABALHO_INDIVIDUO = 'INTERFACE_TRABALHO_INDIVIDUO',
  ORGANIZACAO_DO_TRABALHO_E_CONTEUDO = 'ORGANIZACAO_DO_TRABALHO_E_CONTEUDO',
  PERSONALIDADE = 'PERSONALIDADE',
  RELACOES_SOCIAIS_E_LIDERANCA = 'RELACOES_SOCIAIS_E_LIDERANCA',
  SAUDE_E_BEM_ESTAR = 'SAUDE_E_BEM_ESTAR',
  VALORES_NO_LOCAL_DE_TRABALHO = 'VALORES_NO_LOCAL_DE_TRABALHO',
}