// src/@v2/forms/domain/enums/form-copsoq-dimension.enum.ts
export enum FormCOPSOQDimensionEnum {
  COMPORTAMENTOS_OFENSIVOS = 'COMPORTAMENTOS_OFENSIVOS',
  AMEACAS_DE_VIOLENCIA = 'AMEACAS_DE_VIOLENCIA',
  ASSEDIO_SEXUAL = 'ASSEDIO_SEXUAL',
  ASSEDIO_VIRTUAL = 'ASSEDIO_VIRTUAL',
  ATOS_NEGATIVOS = 'ATOS_NEGATIVOS',
  BULLYING = 'BULLYING',
  CONFLITOS_E_DESENTENDIMENTOS = 'CONFLITOS_E_DESENTENDIMENTOS',
  VIOLENCIA_FISICA = 'VIOLENCIA_FISICA',
  EXIGENCIAS_QUANTITATIVAS = 'EXIGENCIAS_QUANTITATIVAS',
  RITMO_DE_TRABALHO = 'RITMO_DE_TRABALHO',
  DEMANDAS_COGNITIVAS = 'DEMANDAS_COGNITIVAS',
  DEMANDAS_EMOCIONAIS = 'DEMANDAS_EMOCIONAIS',
  DEMANDAS_PARA_ESCONDER_EMOCOES = 'DEMANDAS_PARA_ESCONDER_EMOCOES',
  SATISFACAO_NO_TRABALHO = 'SATISFACAO_NO_TRABALHO',
  INSEGURANCA_NO_TRABALHO = 'INSEGURANCA_NO_TRABALHO',
  CONFLITO_TRABALHO_FAMILIA = 'CONFLITO_TRABALHO_FAMILIA',
  INSEGURANCA_SOBRE_CONDICOES_DE_TRABALHO = 'INSEGURANCA_SOBRE_CONDICOES_DE_TRABALHO',
  QUALIDADE_DO_TRABALHO = 'QUALIDADE_DO_TRABALHO',
  CONTROLE_SOBRE_O_TEMPO_DE_TRABALHO = 'CONTROLE_SOBRE_O_TEMPO_DE_TRABALHO',
  INFLUENCIA_NO_TRABALHO = 'INFLUENCIA_NO_TRABALHO',
  POSSIBILIDADES_DE_DESENVOLVIMENTO = 'POSSIBILIDADES_DE_DESENVOLVIMENTO',
  SIGNIFICADO_DO_TRABALHO = 'SIGNIFICADO_DO_TRABALHO',
  ENGAJAMENTO_NO_TRABALHO = 'ENGAJAMENTO_NO_TRABALHO',
  COMPROMISSO_FACE_AO_LOCAL_DE_TRABALHO = 'COMPROMISSO_FACE_AO_LOCAL_DE_TRABALHO',
  VARIEDADE_DE_TRABALHO = 'VARIEDADE_DE_TRABALHO',
  AUTOEFICACIA = 'AUTOEFICACIA',
  PREVISIBILIDADE = 'PREVISIBILIDADE',
  TRANSPARENCIA_DE_PAPEL = 'TRANSPARENCIA_DE_PAPEL',
  RECOMPENSAS = 'RECOMPENSAS',
  CONFLITOS_LABORAIS = 'CONFLITOS_LABORAIS',
  TAREFAS_ILEGITIMAS = 'TAREFAS_ILEGITIMAS', // Original: Tarefas llegítimas
  APOIO_SOCIAL_DE_COLEGAS = 'APOIO_SOCIAL_DE_COLEGAS',
  APOIO_SOCIAL_DE_SUPERIORES = 'APOIO_SOCIAL_DE_SUPERIORES',
  QUALIDADE_DA_CHEFIA = 'QUALIDADE_DA_CHEFIA',
  SAUDE_GERAL = 'SAUDE_GERAL',
  PROBLEMAS_EM_DORMIR = 'PROBLEMAS_EM_DORMIR',
  BURNOUT = 'BURNOUT',
  ESTRESSE = 'ESTRESSE',
  SINTOMAS_DEPRESSIVOS = 'SINTOMAS_DEPRESSIVOS',
  ESTRESSE_COGNITIVO = 'ESTRESSE_COGNITIVO',
  ESTRESSE_SOMATICO = 'ESTRESSE_SOMATICO',
  SENTIDO_DE_COMUNIDADE_SOCIAL_NO_TRABALHO = 'SENTIDO_DE_COMUNIDADE_SOCIAL_NO_TRABALHO',
  CONFIANCA_HORIZONTAL = 'CONFIANCA_HORIZONTAL',
  CONFIANCA_VERTICAL = 'CONFIANCA_VERTICAL',
  JUSTICA_E_RESPEITO_ORGANIZACIONAL = 'JUSTICA_E_RESPEITO_ORGANIZACIONAL',
}